cleaned kode

+++ nest_server/src/tweety/tweety.service.ts
@@ -1,6 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import mongoose, { Model } from 'mongoose';
 import { CreateTweetyDto } from 'src/tweety/dto/CreateTweetyDto';
 import { Tweety } from 'src/tweety/schema/tweety.schema';
 import { User } from 'src/auth/schema/user.schema'; // uprav cestu podľa projektu;
@@ -14,7 +18,7 @@ export class TweetyService {
   async createTweety(createTweetyDto: CreateTweetyDto) {
     const user = await this.userModel.findById(createTweetyDto.userId);
     if (!user) {
-      throw new Error('User not found');
+      throw new NotFoundException('User not found');
     }
     const newTweety = new this.tweetyModel({
       content: createTweetyDto.content,
@@ -29,12 +33,25 @@ export class TweetyService {
   }
 
   async getTweetyById(id: string) {
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw new BadRequestException('Invalid ID format');
+    }
+
     const tweety = await this.tweetyModel.findById(id);
+    if (!tweety) {
+      throw new NotFoundException('Tweety not found');
+    }
     return tweety;
   }
 
   async deleteTweety(id: string) {
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      throw new BadRequestException('Invalid ID format');
+    }
     const tweety = await this.tweetyModel.findByIdAndDelete(id);
+    if (!tweety) {
+      throw new NotFoundException('Tweety not found');
+    }
     return tweety;
   }
 }
